# apps/backend/docs/swagger.yaml
basePath: /
definitions:
  db.DevicesModel:
    properties:
      createdAt:
        type: string
      deviceType:
        type: string
      endpointArn:
        type: string
      fcmToken:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.NotficationStatus:
    enum:
    - Sent
    - Failed
    - Pending
    type: string
    x-enum-varnames:
    - NotficationStatusSent
    - NotficationStatusFailed
    - NotficationStatusPending
  db.NotificationModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      metadata:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/db.NotficationStatus'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
      website:
        $ref: '#/definitions/db.WebsiteModel'
      websiteId:
        type: string
    type: object
  db.OrganizationMemberModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      organization:
        $ref: '#/definitions/db.OrganizationModel'
      organizationId:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.OrganizationModel:
    properties:
      admin:
        $ref: '#/definitions/db.UserModel'
      adminId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/db.OrganizationMemberModel'
        type: array
      name:
        type: string
      updatedAt:
        type: string
      websites:
        items:
          $ref: '#/definitions/db.WebsiteModel'
        type: array
    type: object
  db.RegionModel:
    properties:
      createdAt:
        type: string
      regionId:
        type: string
      regionName:
        type: string
      ticks:
        items:
          $ref: '#/definitions/db.WebsiteTicksModel'
        type: array
      updatedAt:
        type: string
    type: object
  db.SubscriptionModel:
    properties:
      createdAt:
        type: string
      deviceType:
        type: string
      id:
        type: string
      topicArn:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.UserModel:
    properties:
      createdAt:
        type: string
      devices:
        items:
          $ref: '#/definitions/db.DevicesModel'
        type: array
      email:
        type: string
      id:
        type: string
      notifications:
        items:
          $ref: '#/definitions/db.NotificationModel'
        type: array
      organizationMemberships:
        items:
          $ref: '#/definitions/db.OrganizationMemberModel'
        type: array
      organizationsAsAdmin:
        items:
          $ref: '#/definitions/db.OrganizationModel'
        type: array
      password:
        type: string
      subscriptions:
        items:
          $ref: '#/definitions/db.SubscriptionModel'
        type: array
      updatedAt:
        type: string
      userProfile:
        $ref: '#/definitions/db.UserProfileModel'
    type: object
  db.UserProfileModel:
    properties:
      avatar:
        type: string
      bio:
        type: string
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      phone:
        type: string
      preferences:
        items:
          type: integer
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.WebsiteModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isMonitoringEnabled:
        type: boolean
      notifications:
        items:
          $ref: '#/definitions/db.NotificationModel'
        type: array
      organization:
        $ref: '#/definitions/db.OrganizationModel'
      organizationId:
        type: string
      status:
        $ref: '#/definitions/db.WebsiteStatus'
      ticks:
        items:
          $ref: '#/definitions/db.WebsiteTicksModel'
        type: array
      updatedAt:
        type: string
      url:
        type: string
      websiteName:
        type: string
    type: object
  db.WebsiteStatus:
    enum:
    - Up
    - Down
    - Unknown
    type: string
    x-enum-varnames:
    - WebsiteStatusUp
    - WebsiteStatusDown
    - WebsiteStatusUnknown
  db.WebsiteTicksModel:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      latency:
        type: integer
      region:
        $ref: '#/definitions/db.RegionModel'
      upStatus:
        $ref: '#/definitions/db.WebsiteStatus'
      updatedAt:
        type: string
      website:
        $ref: '#/definitions/db.WebsiteModel'
      websiteId:
        type: string
      websiteRegionId:
        type: string
    type: object
...
  internal_handlers.AuthResponse:
    properties:
      error:
        type: string
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
    type: object
...
  internal_handlers.CreateWebsiteRequest:
    properties:
      organizationId:
        type: string
      url:
        type: string
      websiteName:
        type: string
    type: object
...
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_handlers.SignupRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_handlers.UpdateWebsiteReq:
    properties:
      url:
        type: string
      websiteName:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: My API Name
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
      summary: Log in a user
      tags:
      - auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: Signup request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
      summary: Sign up a new user
      tags:
      - auth
...
  /api/websites:
    get:
      description: Fetches all websites in organizations where user is a member
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.WebsiteModel'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List websites
      tags:
      - websites
    post:
      consumes:
      - application/json
      description: Creates a new website for monitoring in the specified organization
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create website request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateWebsiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.WebsiteModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new website
      tags:
      - websites
  /api/websites/{id}:
    delete:
      description: Removes a website from the system
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.WebsiteModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a website
      tags:
      - websites
    get:
      description: Retrieves a specific website by ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.WebsiteModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a website
      tags:
      - websites
    put:
      consumes:
      - application/json
      description: Updates website name and/or URL
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: Update website request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateWebsiteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.WebsiteModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a website
      tags:
      - websites
...
swagger: "2.0"